fix CI first round
