+++ rdagent/components/coder/data_science/raw_data_loader/eval_tests/data_loader_test.py
@@ -15,9 +15,11 @@ from load_data import load_data
 
 X, y, X_test, test_ids = load_data()
 
+
 # Validate the conditions mentioned in the docstring
 def get_length(data):
     return len(data) if isinstance(data, list) else data.shape[0]
+
 
 assert get_length(X_test) == get_length(test_ids), "Mismatch in length of test images and test IDs"
 assert get_length(X) == get_length(y), "Mismatch in length of training images and labels"
