+++ rdagent/oai/backend/base.py
@@ -1,3 +1,2 @@
-
 class APIBackend:
     """abstract"""
