+++ rdagent/utils/agent/ret.py
@@ -4,8 +4,8 @@ The output of a agent is very important.
 We think this part can be shared.
 """
 
-import re
 import json
+import re
 from abc import abstractclassmethod
 from typing import Any
 
@@ -13,7 +13,8 @@ from rdagent.utils.agent.tpl import T
 
 
 class AgentOut:
-    json_mode: bool = False   # To get the output, is json_mode required.
+    json_mode: bool = False  # To get the output, is json_mode required.
+
     @abstractclassmethod
     def get_spec(cls, **context: Any) -> str:
         raise NotImplementedError(f"Please implement the `get_spec` method")
@@ -34,6 +35,7 @@ class PythonAgentOut(AgentOut):
         if match:
             code = match.group(1)
             return code
+
 
 class BatchEditOut(AgentOut):
     json_mode: bool = True
